Sort it all out
