+++ weyl-frontend/package.json
@@ -93,7 +93,7 @@
       "jsx-a11y/aria-props": 2,
       "jsx-a11y/heading-has-content": 0,
       "jsx-a11y/href-no-hash": 2,
-      "jsx-a11y/label-has-for": 2,
+      "jsx-a11y/label-has-for": 0,
       "jsx-a11y/mouse-events-have-key-events": 2,
       "jsx-a11y/role-has-required-aria-props": 2,
       "jsx-a11y/role-supports-aria-props": 2,
